Added rotation and cube

+++ src/main/Camera.java
@@ -3,7 +3,7 @@ package main;
 public class Camera {
     Vector location;
 
-    public Camera(int x, int y, int z) {
+    public Camera(double x, int y, int z) {
         this.location = new Vector(x, y, z, (byte) 1);
     }
